A length is an `i64`, and the numeric surface is the one the page names
The owner's answer to `docs/open-decisions.md` §2, built, and recorded as
ADR-048.

**A length is an `i64`, and so is an index.** `Vec::len`, `String::len`,
`str::len` and `HashMap::len` - every entry in `std.contracts` that returns a
length. The two conversions are not symmetric, and that is the whole argument:
out of a length cannot fail, since a `usize` exceeds an `i64` only above eight
exabytes of single-byte elements, and the user wrote it in every loop, sum and
comparison that met a length. Into an index can fail, for a negative number -
but a negative index *is* an access out of bounds, which Part III A.2 already
aborts on. Not a new failure mode; the same one a step earlier.

So the trade is a great deal of visible ceremony that cannot fail against one
invisible abort that already exists - and the user writes no conversion at all,
in either direction.

**Both are emitted.** `xs.len()` becomes `xs.len() as i64`. An index becomes
`xs[nikaia_std::index::at(i)]`, where `at` chooses on the type of what is in the
brackets: an integer becomes a `usize`, and anything else - a map's key, most of
all - goes through untouched. A trait and not a function, because this emitter
does not know types (ADR-011 D2) and `counts[path]` is an index too. A negative
index panics as "index out of bounds: the index is -1", which D1 requires: `as
usize` would have said 18446744073709551613, a number nothing in the program
contains.

**An index written only in literals is left alone.** `xs[0]` and `&text[1..3]`
need no conversion and cannot take one: `at(0)` has nothing to infer its argument
type from, and `cannot infer type` about a generated file is what C.1 forbids.

**The machine-width type leaves the writable surface**, and `usize`'s two
`truncating_` entries with it - nothing narrows out of a type nobody can hold.
`check`'s `NUMERIC` is `i32, i64, f64, u8` now; the machine-width types were in it
*because* `len` handed one back. `open-work.md` §2.7's asymmetry is gone with
them, removed rather than answered.

**`u8` is named in Part I 2.2** (D2): `fs::read` already hands back a `Vec[u8]`,
so what was missing was the specification admitting it. `u32`, `u64` and `isize`
stay out until a program asks. The page also now says which types the compiler
accepts without offering, so the list of types is the list of types.

**And the parentheses around a conversion are decided in one place**, which this
forced and ADR-043 D7 needed already. Three things come out as a Rust `as`: a
written cast, `truncating_i32`, and a length. Rust reads such an expression
wrongly in exactly two positions - beside a unary operator, and as the left of a
comparison, where `x as i64 < k` is read as `i64<k>` and the program does not
parse - so parentheses go there and nowhere else. `truncating_` parenthesised
always, which is the other half of C.1: `let n = (big as i32);` is a Rust warning
about a file nobody wrote.

`examples/k-nucleotide.nika` took a `k: usize`, which is a type the surface no
longer offers; it takes an `i64`. What is left is written down in §3 of the
record: a `std` function whose *parameter* is a Rust `usize` still needs the
written conversion, and `json.nika`'s `repeat` is the one in this repository.

+++ crates/nikaia/tests/stored_views.rs
@@ -332,7 +332,7 @@ fn a_view_handed_to_a_call_on_the_subject_is_lowered_too() {
             fn record(&mut self, name: &str, temp: i32) sync {
                 self.stations.insert(name, temp)
             }
-            fn count(&self) -> usize sync { return self.stations.len() }
+            fn count(&self) -> i64 sync { return self.stations.len() }
         }
         fn main() {
             let mut s = Summary::new()
@@ -378,7 +378,7 @@ fn a_view_that_reaches_the_field_through_a_local_is_lowered_too() {
                 let key = name
                 self.stations.insert(key, 1)
             }
-            fn count(&self) -> usize sync { return self.stations.len() }
+            fn count(&self) -> i64 sync { return self.stations.len() }
         }
         fn main() {
             let mut s = Summary::new()
